Update README.md
The web-based system is expected to help common users in making decisions regarding water well in a particular location. The system will be driven by NAQUIM data of CGWB like lithology, geophysical logs, water levels, water quality, aquifer maps, etc. On a user-selected point the AI-based predictor should be able to provide the following information: i. Whether the area is suitable for water well construction or not? ii. At what depth water-bearing zones are expected to be encountered; iii. What is the expected discharge of the well; iv. what is the most suitable drilling technique in the area; v. What is the expected quality of groundwater in the area and other relevant information? There should be a user-friendly graphical user interface. Provisions should also be there to obtain the user feedback in a structured manner.